Fix mobile view and implemented schedule form
