+++ src/layouts/view/MobileView.tsx
@@ -14,10 +14,13 @@ function MobileView({
   theme,
 }: React.PropsWithChildren<Props>) {
   return (
-    <>
+    <div className={'flex flex-col h-screen'}>
       <div className={'h-screen max-md:overflow-y-scroll'}>{children}</div>
-      <MobileNavigationBar items={items} theme={theme} />
-    </>
+
+      <div className={'h-18'}>
+        <MobileNavigationBar items={items} theme={theme} />
+      </div>
+    </div>
   );
 }
